fix old java

+++ src/test/java/fr/dauphine/ja/kiefferachria/MorpionSolitaire/model/Grid5TTest.java
@@ -170,8 +170,8 @@ public class Grid5TTest extends TestCase {
     	}
     	int max=0;
     	for(int i=0;i<grid.getScoreHistory().size();i++) {
-    		if(max<grid.getScoreHistory().get(i))
-    			max=grid.getScoreHistory().get(i);
+    		if(max<grid.getScoreHistory().get(i).intValue())
+    			max=grid.getScoreHistory().get(i).intValue();
     	}
     	System.out.println(max);
 
@@ -195,8 +195,8 @@ public class Grid5TTest extends TestCase {
     	}
     	int max=0;
     	for(int i=0;i<grid.getScoreHistory().size();i++) {
-    		if(max<grid.getScoreHistory().get(i))
-    			max=grid.getScoreHistory().get(i);
+    		if(max<grid.getScoreHistory().get(i).intValue())
+    			max=grid.getScoreHistory().get(i).intValue();
     	}
     	System.out.println(max);
